+++ .rebase/CHANGELOG.md
@@ -2,6 +2,45 @@
 
 The file to keep a list of changed files which will potentionaly help to resolve rebase conflicts.
 
+#### @RomanNikitenko
+https://github.com/che-incubator/che-code/pull/617
+
+- code/src/vs/platform/extensionManagement/common/extensionManagement.ts
+- code/src/vs/platform/extensionManagement/node/extensionManagementService.ts
+- code/src/vs/workbench/contrib/extensions/browser/extensionsWorkbenchService.ts
+- code/src/vs/server/node/serverServices.ts
+---
+
+#### @RomanNikitenko
+https://github.com/che-incubator/che-code/pull/605
+
+- code/src/vs/workbench/browser/web.main.ts
+- code/src/vs/server/node/serverServices.ts
+- code/src/vs/server/node/serverEnvironmentService.ts
+- code/src/vs/workbench/contrib/extensions/browser/extensionsWorkbenchService.ts
+---
+
+#### @sbouchet
+https://github.com/che-incubator/che-code/pull/613
+
+- code/build/package.json
+- code/extensions/microsoft-authentication/package.json
+---
+
+#### @sbouchet
+https://github.com/che-incubator/che-code/pull/621
+
+- code/build/package.json
+- code/test/mcp/package.json
+---
+
+#### @sbouchet
+https://github.com/che-incubator/che-code/pull/611
+
+- code/package.json
+- code/build/npm/gyp/package.json
+---
+
 #### @sbouchet
 https://github.com/che-incubator/che-code/pull/607